Add default role definitions for apps
This PR adds default role definitions for the following applications:

Compliance
Insights
Drift
Vulnerability

Each of these applications needs a base, global read/write permission for
itself, as well as inventory, which is a dependent service which will be
receiving permissions from the upstream service. These permissions are set on
a base role for each application.