Add default role definitions for more apps
This PR adds default role definitions for the following applications:

* ansible-automation
* ansible-hub
* remediations
* subscriptions

Each of these applications needs a base, global read/write permission for
itself, as well as inventory, which is a dependent service which will be
receiving permissions from the upstream service. These permissions are set on
a base role for each application.