Bump `platform_default` role versions
The `platform_default` roles were added to the config _before_ the application
logic to modify the seed script was added to the RBAC repo in: https://github.com/RedHatInsights/insights-rbac/pull/161
Since CI is built off of changes to this repo's `master` branch, those roles
were then added to all CI tenant schemas, _without_ the `platform_default`
flag being set on those roles.

Once the seeding logic went in, since the roles existed and the versions were the
same, the CI schemas did not pick up the `platform_default` role changes, and thus
the platform default groups in CI do not have these roles attached.

This change forces an update to the default roles, to ensure the default groups
pick them up in the seeds.