Remove default roles for apps not integrating with RBAC Lite
In order to not make `system` and/or `platform_default` roles available or enforced
for applications not integrating with the GA rollout of RBAC, we need to remove
these roles from the seed files for the following apps:

- ansible-automation
- ansible-hub
- catalog
- remediations
- sources
- subscriptions

Other apps not needing default roles are:
- inventory
- approval
- sources
- cost-management

However at the time, we don't have default roles for those apps, so we don't need
to take any action.

There will be a PR added to the [RBAC app repo](https://github.com/RedHatInsights/insights-rbac/)
in order to actually remove these roles from existing tenant schemas.

This PR is to remove the roles from the seeds to prevent further creation of these
default roles.