Add GitHub action and permission validation
This sets up a GitHub action/workflow to run two jobs for validating our
permission and role JSON. Any permission/role file that's updated in a PR will
be validated against the appropriate schema.

A schema has been added to validate the structure of our permission files.

A second commit will add the same for roles.