Create a PR on merge to `master`
When a commit is pushed/merged into `master`, we'll create a PR based on the outcome
of what's pushed to the `configmaps` branch, which will be the generated ConfigMaps
from our `rbac-config-actions` run.

We have to do this currently, because our `master` branch requires PR review for
any commits, so we can't directly commit from our action to `master`.