Bump RBAC role in stage for Kafka testing
We experienced a kafka issue in stage on Monday which we need to confirm is resolved.
In order to trigger the behavior, we need to run seeds which update a role. This
will force an update on the "User Access principal viewer" role, which will
attempt to send a kafka message.