feature: remaining permissions for Notifications
These temporary permissions will help us deal with the system data like
bundles, applications and event types, and the tenant-specific data like
behavior groups.

In the future we will refactor the system data calls in the
Notifications code to comply with the Kessel style, by querying for the
specific elements instead of querying for the broad workspace
permissions.

RHCLOUD-35813